Extract presentational ToggleButtonView from ToggleButton
