+++ src/components/toggle-button/ToggleButton.tsx
@@ -1,37 +1,15 @@
 import { useRef, useCallback, useEffect, useContext } from "react";
-import { gsap } from "gsap";
 import { SongContext } from "../../contexts/contexts";
 import { TestingContext } from "../../contexts/contexts";
-import { LayoutContext } from "../../contexts/contexts";
 import { WebAudioContext } from "../../contexts/contexts";
 import { useMusicPlayerStore } from "../../stores/musicPlayerStore";
 import { nextSubdivision } from "../../utils/audioUtils";
-import "../../styles/components/Icon.css";
-import "../../styles/components/ToggleButton.css";
-
-const START_PARAMS = {
-  rotateZ: "-180",
-  backgroundColor: "rgba(255, 255, 255, .3)",
-  points:
-    "6.69872981 6.69872981 93.01270188 6.69872981 93.01270188 50 93.01270188 93.01270188 6.69872981 93.01270188",
-};
-
-const STOP_PARAMS = {
-  rotateZ: "0",
-  backgroundColor: "rgba(255, 255, 255, 0)",
-  points:
-    "6.69872981 0 6.69872981 0 93.01270188 50 6.69872981 100 6.69872981 100",
-};
+import {
+  ToggleButtonView,
+  ToggleButtonViewHandle,
+} from "./ToggleButtonView";
 
 type PlayerState = "stopped" | "pending-start" | "active" | "pending-stop";
-
-interface AnimationTargets {
-  button: HTMLButtonElement;
-  circleSvg: Element;
-  iconDiv: Element;
-  iconSvg: Element;
-  iconPoly: Element;
-}
 
 interface Props {
   name: string;
@@ -41,12 +19,10 @@ interface Props {
 }
 
 export const ToggleButton = (props: Props) => {
-  const buttonRef = useRef<HTMLButtonElement>(null);
-  const animationTargetsRef = useRef<AnimationTargets | null>(null);
+  const viewRef = useRef<ToggleButtonViewHandle>(null);
   const animationEventRef = useRef<number | undefined>(undefined);
 
   const { WAW } = useContext(WebAudioContext)!;
-  const { vh } = useContext(LayoutContext)!;
   const { id, timeSignature, bpm } = useContext(SongContext)!;
   const { flags } = useContext(TestingContext)!;
   const { name, groupName, quantizeLength } = props;
@@ -69,79 +45,9 @@ export const ToggleButton = (props: Props) => {
   const quantizedStartBeats = flags.quantizeSamples
     ? timeSignature * parseInt(quantizeLength!)
     : 1;
-  const buttonRadius = vh ? vh * 3.5 : 0;
-  const buttonBorder = vh ? (vh * 3.5) / 15 : 0;
 
   const changePlayerState = useCallback(
     (newState: PlayerState) => {
-      const runAnimation = (type: "start" | "stop", duration: number) => {
-        const seconds = duration / 1000;
-        const targets = animationTargetsRef.current!;
-
-        // clear queue
-        gsap.killTweensOf(targets.circleSvg);
-        gsap.killTweensOf(targets.iconPoly);
-        gsap.killTweensOf(targets.iconDiv);
-        gsap.killTweensOf([...targets.iconDiv.children]);
-        gsap.killTweensOf(targets.button);
-
-        let points: string | undefined;
-        let backgroundColor: string | undefined;
-        let rotateZ: number;
-
-        if (type === "start") {
-          rotateZ = -180;
-          backgroundColor = START_PARAMS.backgroundColor;
-          points = START_PARAMS.points;
-
-          // sweep always begins from a 0 offset regardless of current value
-          gsap.fromTo(
-            targets.circleSvg,
-            { strokeDashoffset: 0 },
-            {
-              strokeDashoffset: 2 * Math.PI * (buttonRadius - buttonBorder / 2),
-              duration: seconds,
-              ease: "none",
-            }
-          );
-        } else {
-          rotateZ = 0;
-          backgroundColor = STOP_PARAMS.backgroundColor;
-          points = STOP_PARAMS.points;
-
-          // run circle animation
-          gsap.to(targets.circleSvg, {
-            strokeDashoffset: 0,
-            duration: seconds,
-            ease: "none",
-          });
-        }
-
-        // run icon animation (morph polygon points)
-        gsap.to(targets.iconPoly, {
-          attr: { points },
-          duration: seconds,
-          ease: "none",
-        });
-
-        // run rotate animation
-        gsap.to(
-          [targets.iconDiv, ...targets.iconDiv.children],
-          {
-            rotation: rotateZ,
-            duration: seconds,
-            ease: "none",
-          }
-        );
-
-        // run button animation
-        gsap.to(targets.button, {
-          backgroundColor,
-          duration: seconds,
-          ease: "power2.in",
-        });
-      };
-
       // cancel current event for this toggle (necessary to stop a pending start)
       scheduler.cancel(animationEventRef.current);
 
@@ -181,7 +87,7 @@ export const ToggleButton = (props: Props) => {
       const quantizedStartMillis =
         (quantizedStartSeconds - audioCtx.currentTime) * 1000;
       const animationType = newState === "stopped" ? "stop" : "start";
-      runAnimation(animationType, quantizedStartMillis);
+      viewRef.current!.runAnimation(animationType, quantizedStartMillis);
     },
     [
       scheduler,
@@ -189,8 +95,6 @@ export const ToggleButton = (props: Props) => {
       audioCtx,
       bpm,
       quantizedStartBeats,
-      buttonRadius,
-      buttonBorder,
       player,
       updateVoiceState,
       queueVoice,
@@ -200,21 +104,11 @@ export const ToggleButton = (props: Props) => {
 
   /* Initialize Hook */
   useEffect(() => {
-    const btn = buttonRef.current!;
-    // store the animation targets based on their relative positions in the DOM
-    animationTargetsRef.current = {
-      button: btn,
-      circleSvg: btn.children[0],
-      iconDiv: btn.children[1],
-      iconSvg: btn.children[1].children[0],
-      iconPoly: btn.children[1].children[0].children[0],
-    };
-
     addVoice({
       id: props.name,
       group: props.groupName,
       voiceState: "stopped",
-      ref: btn,
+      ref: viewRef.current!.getButton()!,
     });
   }, [id, addVoice, props.groupName, props.name]);
 
@@ -241,9 +135,9 @@ export const ToggleButton = (props: Props) => {
   }, [player]);
 
   return (
-    <button
-      className="toggle-button"
-      ref={buttonRef}
+    <ToggleButtonView
+      ref={viewRef}
+      active={playerState === "active"}
       onClick={() => {
         switch (playerState) {
           case "stopped": // start if stopped
@@ -261,49 +155,6 @@ export const ToggleButton = (props: Props) => {
             break;
         }
       }}
-      style={{
-        cursor: "pointer",
-        height: buttonRadius * 2,
-        width: buttonRadius * 2,
-      }}
-    >
-      <svg
-        className="svg"
-        width={2 * buttonRadius}
-        height={2 * buttonRadius}
-        style={{
-          strokeDashoffset:
-            playerState === "active"
-              ? 2 * Math.PI * (buttonRadius - buttonBorder / 2)
-              : 0,
-        }}
-      >
-        <circle
-          className="svg-circle"
-          cx={buttonRadius}
-          cy={buttonRadius}
-          r={buttonRadius - buttonBorder / 2}
-          style={{
-            strokeWidth: buttonBorder,
-            strokeDasharray: 2 * Math.PI * (buttonRadius - buttonBorder / 2),
-          }}
-        />
-      </svg>
-
-      <div className={`scale-div-morph toggle-icon`}>
-        <svg
-          viewBox="0 0 100 100"
-          xmlns="http://www.w3.org/2000/svg"
-          xmlnsXlink="http://www.w3.org/1999/xlink"
-          className={`toggle-icon icon-white`}
-        >
-          <polygon
-            id="icon-play3-poly"
-            className={`icon icon-white`}
-            points={STOP_PARAMS.points}
-          />
-        </svg>
-      </div>
-    </button>
+    />
   );
 };
